refactor(FramesTest): BowlingGameTest.java를 FramesTest로 이름 변경
